git-svn-id: http://192.168.0.243:4435/svn/bee/branches/hoptransa_logistics@2265 aaee961a-53e7-2249-bd01-b2a8f0c4ce08

+++ src/com/butent/bee/client/modules/ec/EcWidgetFactory.java
@@ -15,7 +15,7 @@ public final class EcWidgetFactory {
   private static final String STYLE_FIELD_CONTAINER = "-container";
   private static final String STYLE_FIELD_LABEL = "-label";
 
-  public static Widget createStockWidget(int stock) {
+  public static Widget createStockWidget(double stock) {
     InlineLabel widget = new InlineLabel();
     widget.getElement().setInnerText(EcKeeper.formatStock(stock));
