+++ src/com/butent/bee/client/modules/ec/widget/ItemDetails.java
@@ -285,7 +285,7 @@ public class ItemDetails extends Flow {
             Label warehouseWidget = new Label(EcKeeper.getWarehouseLabel(warehouse));
             table.setWidgetAndStyle(row, 0, warehouseWidget, stylePrefix + "warehouse");
 
-            int remainder = BeeUtils.toInt(as.getRemainders().get(warehouse));
+            double remainder = BeeUtils.toDouble(as.getRemainders().get(warehouse));
             Widget stockWidget = EcWidgetFactory.createStockWidget(remainder);
             table.setWidgetAndStyle(row, 1, stockWidget, stylePrefix + "stock");
